+++ premium/backend/src/baserow_premium/api/row_comments/views.py
@@ -1,6 +1,7 @@
 from django.conf import settings
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import extend_schema, OpenApiParameter
+from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -34,10 +35,32 @@ class RowCommentView(APIView):
                 description="The row to get row comments for.",
             ),
             OpenApiParameter(
+                name="limit",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Defines how many rows should be returned.",
+            ),
+            OpenApiParameter(
+                name="offset",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Can only be used in combination with the `limit` "
+                "parameter and defines from which offset the rows should "
+                "be returned.",
+            ),
+            OpenApiParameter(
                 name="page",
                 location=OpenApiParameter.QUERY,
                 type=OpenApiTypes.INT,
-                description="Selects which page of row comments should be returned.",
+                description="Defines which page of rows should be returned. Either "
+                "the `page` or `limit` can be provided, not both.",
+            ),
+            OpenApiParameter(
+                name="size",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Can only be used in combination with the `page` parameter "
+                "and defines how many rows should be returned.",
             ),
         ],
         tags=["Database table rows"],
@@ -64,12 +87,17 @@ class RowCommentView(APIView):
     def get(self, request, table_id, row_id):
         comments = RowCommentHandler.get_comments(request.user, table_id, row_id)
 
-        paginator = PageNumberPagination(
-            limit_page_size=settings.ROW_COMMENT_PAGE_SIZE_LIMIT
-        )
+        if LimitOffsetPagination.limit_query_param in request.GET:
+            paginator = LimitOffsetPagination()
+            paginator.max_limit = settings.ROW_COMMENT_PAGE_SIZE_LIMIT
+        else:
+            paginator = PageNumberPagination(
+                limit_page_size=settings.ROW_COMMENT_PAGE_SIZE_LIMIT
+            )
 
         page = paginator.paginate_queryset(comments, request, self)
-        serializer = RowCommentSerializer(page, many=True)
+        context = {"user": request.user}
+        serializer = RowCommentSerializer(page, many=True, context=context)
 
         return paginator.get_paginated_response(serializer.data)
 
@@ -115,4 +143,5 @@ class RowCommentView(APIView):
         new_row_comment = RowCommentHandler.create_comment(
             request.user, table_id, row_id, data["comment"]
         )
-        return Response(RowCommentSerializer(new_row_comment).data)
+        context = {"user": request.user}
+        return Response(RowCommentSerializer(new_row_comment, context=context).data)